reformating code to be clearer - Part1: AppProcesses + Package split
