+++ src/main/java/main/UI/menu/GraphicalMenus.java
@@ -1,0 +1,8 @@
+package main.UI.menu;
+
+public class GraphicalMenus {
+
+    HomeScreen homeScreen;
+    OptionsMenu optionsMenu;
+    QuickMenu quickMenu;
+}
