+++ src/main/java/main/gaze/devicemanager/TobiiGazeDeviceManager.java
@@ -2,7 +2,6 @@ package main.gaze.devicemanager;
 
 import lombok.extern.slf4j.Slf4j;
 import main.Configuration;
-import main.SecondStage;
 
 import java.awt.*;
 import java.util.concurrent.ExecutorService;
