+++ src/main/java/main/process/GazePlayProcess.java
@@ -1,23 +1,16 @@
 package main.process;
 
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.BorderPane;
-import main.ProgressButton;
-import main.SecondStage;
-import main.UtilsOS;
-import main.gaze.devicemanager.AbstractGazeDeviceManager;
+import main.utils.UtilsOS;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.LinkedList;
 
 public class GazePlayProcess implements AppProcess {
-    ProcessBuilder pb;
+    ProcessBuilder processBuilder;
     String gazePlayInstallationRepo;
 
-    public GazePlayProcess( String gazePlayInstallationRepo){
+    public GazePlayProcess(String gazePlayInstallationRepo) {
         super();
         this.gazePlayInstallationRepo = gazePlayInstallationRepo;
     }
@@ -25,7 +18,7 @@ public class GazePlayProcess implements AppProcess {
     @Override
     public Process start() {
         try {
-            return pb.inheritIO().start();
+            return processBuilder.inheritIO().start();
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -33,13 +26,13 @@ public class GazePlayProcess implements AppProcess {
     }
 
     @Override
-    public void init() {
-        pb = createGazePlayBuilder();
+    public void setUpProcessBuilder() {
+        processBuilder = createGazePlayBuilder();
     }
 
     private ProcessBuilder createGazePlayBuilder() {
         String javaBin;
-        if(UtilsOS.isWindows()) {
+        if (UtilsOS.isWindows()) {
             javaBin = gazePlayInstallationRepo + "/lib/jre/bin/java.exe";
         } else {
             javaBin = gazePlayInstallationRepo + "/lib/jre/bin/java";
@@ -49,25 +42,6 @@ public class GazePlayProcess implements AppProcess {
         LinkedList<String> commands = new LinkedList<>(Arrays.asList(javaBin, "-cp", classpath, "net.gazeplay.GazePlayLauncher"));
 
         return new ProcessBuilder(commands);
-    }
-
-    @Override
-    public ProgressButton createButton(BorderPane borderPane, SecondStage stage, AbstractGazeDeviceManager tgdm) {
-        ProgressButton pb = new ProgressButton();
-        ImageView logo = new ImageView(new Image("images/gazeplayicon.png"));
-        pb.getButton().setRadius(100);
-        logo.setFitWidth(pb.getButton().getRadius() * 0.7);
-        logo.setFitHeight(pb.getButton().getRadius() * 0.7);
-        logo.fitWidthProperty().bind(pb.getButton().radiusProperty().multiply(0.7));
-        logo.fitHeightProperty().bind(pb.getButton().radiusProperty().multiply(0.7));
-        logo.setPreserveRatio(true);
-        pb.setImage(logo);
-        pb.assignIndicator((e) -> {
-            stage.proc = this.start();
-        }, 500);
-        this.init();
-        pb.active();
-        return pb;
     }
 
 }
