+++ src/main/java/main/utils/UtilsOS.java
@@ -1,4 +1,4 @@
-package main;
+package main.utils;
 
 public class UtilsOS {
